Add back TEXT_OFFSET in the built image...
Some SDM845-based phones, namely the Pixel 3 (blueline) will fail to
boot a recent kernel on the Q or R bootloader if this is missing.

Meanwhile, the P bootloader can deal with that image properly it seems.

With Q:

```
Booting from slot (a)
KernelDecompress failed: Invalid Parameter Kernel Load Address unaligned
```

With R:

```
Booting from slot (a)
Partition Not Found
KernelDecompress failed: Invalid Parameter Kernel TextOffset does not match
```

This is theorized to be required for other SDM845 devices, but untested.

This (hackily) reverts parts of
120dc60d0bdbadcad7460222f74c9ed15cdeb73e.